WIP on LayerListPane
