+++ weyl-frontend/app/components/LayerListPane/index.js
@@ -1,7 +1,10 @@
 import React from "react";
+import {
+  Classes,
+} from "@blueprintjs/core";
+import * as _ from "underscore";
 import LayerListItem from "./LayerListItem";
 import Pane from "../Pane";
-const _ = require("underscore");
 
 
 class LayerListPane extends React.Component { // eslint-disable-line react/prefer-stateless-function
@@ -11,6 +14,36 @@ class LayerListPane extends React.Component { // eslint-disable-line react/prefe
       activeLayerId: null,
       activeMode: null,
     };
+  }
+
+  render() {
+    return (
+      <Pane
+        title="Layers"
+        iconName="layers"
+        visible={this.props.visible && _.size(this.props.layers.toArray()) > 0}
+        onClose={this.props.onClose}
+      >
+        <table className="pt-table pt-striped pt-condensed" style={{ width: "100%" }}>
+          <tbody>
+            {
+              _.map(this.props.layers.toArray(), layer =>
+                <LayerListItem
+                  key={layer.get("id")}
+                  layer={layer.toJS()}
+                  onButtonClick={(name) => this.onButtonClick(name, layer.get("id"))}
+                  onToggleValueVisible={this.props.onToggleValueVisible}
+                  onLayerStyleChange={(attribute) => this.props.onLayerStyleChange(layer.get("id"), "ATTRIBUTE", attribute)}
+                  onLayerThemeChange={(idx) => this.props.onLayerStyleChange(layer.get("id"), "THEME", idx)}
+                  onBufferClick={(bufferDistance) => this.onBufferCompute(layer.get("id"), bufferDistance)}
+                  mode={(this.state.activeLayerId === layer.get("id")) ? this.state.activeMode : null}
+                />
+              )
+            }
+          </tbody>
+        </table>
+      </Pane>
+    );
   }
 
   onButtonClick(name, layerId) {
@@ -33,37 +66,6 @@ class LayerListPane extends React.Component { // eslint-disable-line react/prefe
       layerId,
       bufferDistance,
     });
-  }
-
-  render() {
-    const rows = [];
-    this.props.layers.toArray().forEach(layer =>
-      rows.push(
-        <LayerListItem
-          key={layer.get("id")}
-          layer={layer.toJS()}
-          onButtonClick={(name) => this.onButtonClick(name, layer.get("id"))}
-          onToggleValueVisible={this.props.onToggleValueVisible}
-          onLayerStyleChange={(attribute) => this.props.onLayerStyleChange(layer.get("id"), "ATTRIBUTE", attribute)}
-          onLayerThemeChange={(idx) => this.props.onLayerStyleChange(layer.get("id"), "THEME", idx)}
-          onBufferClick={(bufferDistance) => this.onBufferCompute(layer.get("id"), bufferDistance)}
-          mode={(this.state.activeLayerId === layer.get("id")) ? this.state.activeMode : null}
-        />
-      )
-    );
-
-    return (
-      <Pane
-        title="Layers"
-        iconName="layers"
-        visible={this.props.visible && _.size(this.props.layers.toArray()) > 0}
-        onClose={this.props.onClose}
-      >
-        <div className="ui divided items">
-          {rows}
-        </div>
-      </Pane>
-    );
   }
 }
