+++ weyl-frontend/app/components/LayerListPane/styles.css
@@ -2,7 +2,7 @@
   padding-top: 10px;
 }
 
-a {
+/*a {
   color: black;
   opacity: 0.6;
 }
@@ -18,4 +18,4 @@ a.active {
 
 a.enabled {
   color: blue;
-}
+}*/
